refactor: deduplicate LLM response test helpers into shared module
Extracts _make_text_response, _make_tool_call_response, _make_tool_call,
_make_mixed_response, _make_null_content_tool_call_response, and
_make_tool_call_malformed_args into tests/llm_response_fixtures.py.
Updates nine test files to import from there and drops local copies.

Closes #27
