+++ tests/test_agent_loop_plugin.py
@@ -19,6 +19,11 @@ from corvidae.persistence import PersistencePlugin
 from corvidae.tool import ToolRegistry
 
 from helpers import build_plugin_and_channel, drain
+from llm_response_fixtures import (
+    _make_text_response,
+    _make_tool_call_response,
+    _make_tool_call,
+)
 
 
 # ---------------------------------------------------------------------------
@@ -44,34 +49,6 @@ AGENT_DEFAULTS = {
     "max_context_tokens": 8000,
     "keep_thinking_in_history": False,
 }
-
-
-def _make_text_response(text: str) -> dict:
-    return {"choices": [{"message": {"role": "assistant", "content": text}}]}
-
-
-def _make_tool_call_response(calls: list[dict]) -> dict:
-    return {
-        "choices": [
-            {
-                "message": {
-                    "role": "assistant",
-                    "content": "",
-                    "tool_calls": calls,
-                }
-            }
-        ]
-    }
-
-
-def _make_tool_call(call_id: str, name: str, args: dict) -> dict:
-    return {
-        "id": call_id,
-        "function": {
-            "name": name,
-            "arguments": json.dumps(args),
-        },
-    }
 
 
 # ---------------------------------------------------------------------------
